+++ src/test/kotlin/ru/yoomoney/gradle/plugins/moira/trigger/MoiraTriggerPluginSpec.kt
@@ -11,7 +11,7 @@ class MoiraTriggerPluginSpec : AbstractPluginSpec() {
         buildFile.appendText("""
 
             dependencies {
-                moiraFromDirCompile 'com.yandex.money.tech:moira-kotlin-dsl:1.0.3'
+                moiraFromDirCompile 'ru.yoomoney.tech:moira-kotlin-dsl:1.4.1-feature-github-SNAPSHOT'
             }
 
         """.trimIndent())
@@ -27,7 +27,7 @@ class MoiraTriggerPluginSpec : AbstractPluginSpec() {
         buildFile.appendText("""
 
             dependencies {
-                moiraFromDirCompile 'com.yandex.money.tech:moira-kotlin-dsl:1.0.3'
+                moiraFromDirCompile 'ru.yoomoney.tech:moira-kotlin-dsl:1.4.1-feature-github-SNAPSHOT'
             }
 
         """.trimIndent())