+++ src/test/resources/ru/yoomoney/gradle/plugins/moira/trigger/collect/SingleTrigger.kts
@@ -1,8 +1,8 @@
 package ru.yoomoney.gradle.plugins.moira.trigger.collect
 
-import ru.yandex.money.moira.dsl.triggers.TriggerState.ERROR
-import ru.yandex.money.moira.dsl.triggers.TriggerState.OK
-import ru.yandex.money.moira.dsl.triggers.trigger
+import ru.yoomoney.tech.moira.dsl.triggers.TriggerState.ERROR
+import ru.yoomoney.tech.moira.dsl.triggers.TriggerState.OK
+import ru.yoomoney.tech.moira.dsl.triggers.trigger
 import java.time.Duration
 
 trigger(id = "test_trigger", name = "Test trigger") {