+++ src/test/resources/ru/yoomoney/gradle/plugins/moira/trigger/collect/TriggersList.kts
@@ -1,8 +1,8 @@
 package ru.yoomoney.gradle.plugins.moira.trigger.collect
 
-import ru.yandex.money.moira.dsl.triggers.TriggerState.ERROR
-import ru.yandex.money.moira.dsl.triggers.TriggerState.OK
-import ru.yandex.money.moira.dsl.triggers.triggers
+import ru.yoomoney.tech.moira.dsl.triggers.TriggerState.ERROR
+import ru.yoomoney.tech.moira.dsl.triggers.TriggerState.OK
+import ru.yoomoney.tech.moira.dsl.triggers.triggers
 import java.time.Duration
 
 triggers {