+++ src/test/resources/ru/yoomoney/gradle/plugins/moira/trigger/collect/test/throw/ThrowExceptionTrigger.kts
@@ -1,6 +1,6 @@
 package ru.yoomoney.gradle.plugins.moira.trigger.collect.test.`throw`
 
-import ru.yandex.money.moira.dsl.triggers.trigger
+import ru.yoomoney.tech.moira.dsl.triggers.trigger
 
 trigger(id = "throw_exception_test_trigger", name = "Throw exception test trigger") {
     description = "Trigger for tests only"